+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #113 Upgrade to Jetty 10.0.13 / 11.0.13
 * #115 JettyTester - support for testing redirects
 * #114 Upgrade of Jetty to 10.0.14 / 11.0.14
+* #117 JettyTester's ResponseMatcher must allow access to any type of content
 
 ## 3.0.M1